Fix index types

+++ luciferase/io.py
@@ -160,14 +160,15 @@ class TecanReader:
                     .replace('OVER', np.inf) \
                     .apply(pd.to_numeric, errors='coerce')
             times = pd.Index(pd.to_timedelta(raw.iloc[0], unit='S')).rename('time')
-            temps = pd.Index(raw.iloc[1]).rename('temperature')
+            temps = pd.Index(raw.iloc[1]).rename('temperature').astype('float')
             data = raw.rename_axis(index='well', columns=raw.index.name) \
                       .drop(raw[~raw.index.to_series().astype('str').str.match("[A-Z]{1,2}[0-9]{1,3}")].index)
             index = data.index.to_series().str.extract('([A-Z]+)([0-9]+)') \
                         .rename({0: 'row', 1: 'column'}, axis='columns').apply(pd.to_numeric, errors='ignore') \
                         .set_index(['row', 'column']).index
             columns = pd.MultiIndex.from_arrays(
-                [data.columns.rename('cycle'), pd.Series([name for _ in data.columns], name='label'), times, temps])
+                [data.columns.rename('cycle').astype('int'),
+                 pd.Series([name for _ in data.columns], name='label'), times, temps])
             data = pd.DataFrame(data.values, index=index, columns=columns) \
                      .drop(columns[np.isnat(columns.get_level_values(2))], axis='columns')
         elif cell.value == "<>":
